+++ source/Relativity.Testing.Framework.Api/Strategies/MOTD/IMotdUpdateStrategy.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+using Relativity.Testing.Framework.Models;
+
+namespace Relativity.Testing.Framework.Api.Strategies
+{
+	internal interface IMotdUpdateStrategy
+	{
+		MessageOfTheDay Update(MessageOfTheDay entity);
+
+		Task<MessageOfTheDay> UpdateAsync(MessageOfTheDay entity);
+	}
+}