+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.21.0] - 2021-07-23
+
+### Added
+
+- IsTextOnly method in IMessageOfTheDayService. - [RTF-1260](https://jira.kcura.com/browse/RTF-1260)
+- Support for V1 Notifications Manager (MotD) API which will be used for Relativity 12.1+ - [RTF-1260](https://jira.kcura.com/browse/RTF-1260)
+
 ## [3.20.0] - 2021-07-22
 
 ### Changed